Moved event types to events.go file.
